pyproject.toml
[project]
name = "gitatp"
version = "0.4.0"
description = "Git over ATProto"
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    { name = "Public Domain", email = "publicdomainrelay@protonmail.com" }
]
license = {text = "Unlicense"}
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.55",
    "snoop>=0.6.0",
    "pydantic>=2.10.2",
    "keyring>=25.5.0",
    "pygments>=2.18.0",
    "markdown2>=2.5.1",
    "fastapi>=0.115.5",
    "fastui>=0.7.0",
    "aiohttp-asgi>=0.5.2",
    "python-magic>=0.4.27",
]

[project.urls]
Repository = "https://github.com/publicdomainrelay/gitatp.git"
Issues = "https://github.com/publicdomainrelay/gitatp/issues"

[project.scripts]
gitatp = "gitatp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
